requirements.txt
pytest==8.3.4
pytest-asyncio==0.25.3
allure-pytest==2.13.5
playwright==1.41.0
requests==2.32.3